+++ test/test_index.dart
@@ -8,7 +8,7 @@ part of 'text_indexing_test.dart';
 /// read/write operations to the [dictionary] and [postings].
 ///
 /// Use for testing and examples.
-class _TestIndex {
+class _TestIndex with InvertedIndexMixin implements InvertedIndex {
   //
 
   /// The [Dictionary] instance that is the data-store for the index's term
@@ -24,6 +24,7 @@ class _TestIndex {
   /// Returns a subset of [postings] corresponding to [terms].
   ///
   /// Simulates latency of 50 milliseconds.
+  @override
   Future<Postings> getPostings(Iterable<String> terms) async {
     final Postings retVal = {};
     for (final term in terms) {
@@ -40,6 +41,7 @@ class _TestIndex {
   /// Adds/overwrites the [values] to [dictionary].
   ///
   /// Simulates latency of 50 milliseconds.
+  @override
   Future<void> upsertDictionary(Dictionary values) async {
     /// Simulate write latency of 50milliseconds.
     await Future.delayed(const Duration(milliseconds: 50));
@@ -51,7 +53,10 @@ class _TestIndex {
   /// Adds/overwrites the [values] to [postings].
   ///
   /// Simulates latency of 50 milliseconds.
+  @override
   Future<void> upsertPostings(Postings values) async {
+    /// Simulate write latency of 50milliseconds.
+    await Future.delayed(const Duration(milliseconds: 50));
     postings.addAll(values);
   }
 
@@ -60,6 +65,7 @@ class _TestIndex {
   /// Returns a subset of [dictionary] corresponding to [terms].
   ///
   /// Simulates latency of 50 milliseconds.
+  @override
   Future<Dictionary> getDictionary([Iterable<String>? terms]) async {
     if (terms == null) return dictionary;
     final Dictionary retVal = {};
@@ -71,4 +77,17 @@ class _TestIndex {
     }
     return retVal;
   }
+
+  @override
+  final ITextAnalyzer analyzer = TextAnalyzer();
+
+  @override
+  Future<Ft> get vocabularyLength async {
+    /// Simulate write latency of 50milliseconds.
+    await Future.delayed(const Duration(milliseconds: 50));
+    return dictionary.length;
+  }
+
+  @override
+  int get phraseLength => 3;
 }